Put client card creation in another file.
